receiveAttack first test passed.
Gameboard has a function that deploy 10 ships using the ship() function and saves the coordinates, because there isn't much logic and i already tested ship() i didn't tested deployShips() because it would be redundant.
receiveAttack gets a coordinate, and check if this coordinate is already occupied by any of the ships, if it is, it runs a hit() on the specific ship, and also saves the coordinates just in case the attack coordinates didn't hit any ship.