Second DOMLogic test passed.
"can click a specific coordinate in the cpu board and run receiveAttack"

tests 11 and 12 will be fixed on the next commit, because they don't affect directly the game logic.
they are broken because i needed to change a function to a constant in order to change the original ships object, instead of creating a new instance, this fix allowed me to click on a specific coordinate, and the ship with that coordinate actually receives the hit.