re-factorized missedAttacks into receivedAttacks.
missedAttacks wasn't a needed property, i changed it because i needed to make a turn algorithm in order to make and receive attacks from the cpu.
If i kept missedAttacks, i would also need successAttacks to know if i need to switch turns.
that's why i re-factorized that property into receivedAttacks, which will be much useful as information to make the turns algorithm.