fixed attackCorners on cpuAttacks + other tiny fix
because the attackCorners logic was doing attacks directly on the player's gameBoard, the attackPlayer algorithm inside of the cpuObj, didn't knew about these made attacks, so it was repeating these attacked coords, wasting turns.
so i had to call the attackPlayer function inside of the attackCorners to make the corners attacks, also made that these corners attacks can be repeated, adding a "ignoreCoords" parameter.

also fixed the number range of the generated coords inside of attackPlayer, the cpu wasn't attacking the outer layers.