re-worked attack arrays inside of cpuObj.
it wasn't necessary to have a usedCoords array and push the done attacks into it, we already had the playerObj to get the attacks that it had received.
Also, this playerObj received attacks array, had the coords attacked when a ship is sunk, the old usedCoords didn't had it, so now we have don't have any chance of a repeated hit.
With this better info, we can also build up the logic on the ia improvements a lot easier.